january 4th 2020

+++ TEP/src/nested/Leap_Year.java
@@ -5,13 +5,13 @@ import java.util.Scanner;
 public class Leap_Year {
 
 	public static void main(String[] args) {
-		int year;
+		byte year;
 
 		Scanner scanner=new Scanner (System.in);
 
 		System.out.println(" Enter a year");
 
-		year= scanner.nextInt();
+		year= (byte) scanner.nextInt();
 
 		if (year%4 ==0 && year%100!=0)
